Now compiles with g++

+++ source/library/Posiform.h
@@ -464,7 +464,7 @@ namespace Petter
 		real energy = packing.solve_lp(sol);
 		energy += constant + packingconst;
 
-		for (size_t i=0;i<size_t(n);++i) {
+		for (int i=0;i<n;++i) {
 			if (var_used[i] || x.at(i)<0) {
 				x.at(i) = sol.at(i);
 			}
